pyproject.toml
[tool.poetry]
name = "modern-colorthief"
version = "0.1.0"
description = ""
authors = ["baseplate-admin <61817579+baseplate-admin@users.noreply.github.com>"]
readme = "README.md"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
packages = [{include = "modern_colorthief"}]

[tool.poetry.dependencies]
python = "^3.8"
pillow = "^10.1.0"


[tool.poetry.group.test.dependencies]
pytest = "^6.0.0"
poethepoet = "^0.24.2"
colorthief = "^0.2.1"
pillow-avif-plugin = "^1.4.1"

[tool.poetry.group.dev.dependencies]
black = "^23.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
